// SentimentAnalyser/ClientApp/src/generated/api-client/model/analyzeTextResponse.ts
/**
 * Sentiment Analyser API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AnalyzeTextResponse { 
    html?: string | null;
    score?: number;
}